+++ lib/src/wistia_player_flags.dart
@@ -1,6 +1,0 @@
-class WistiaPlayerFlags {
-  final bool isMuted;
-  final bool forceHD;
-
-  const WistiaPlayerFlags({this.isMuted = false, this.forceHD = false});
-}