+++ lib/wistia_player.dart
@@ -2,3 +2,7 @@ library wistia_player;
 
 export 'src/wistia_player.dart';
 export 'src/wistia_player_controller.dart';
+export 'src/wistia_player_options.dart';
+export 'src/enums/end_video_behavior.dart';
+export 'src/enums/preload.dart';
+export 'src/enums/wistia_player_state.dart';