+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 ## 0.0.1
 
 * Initial release
+
+## 0.0.2
+
+* Add more default value for `WistiaPlayerValue`
+
+## 0.0.3
+
+* Refactor options passing to be passed to the Wistia div class
+* Add `convertUrlToId` to convert Wistia shareable URL to Wistia ID to pass to `WistiaPlayerController` videoId property
+* Add `WistiaPlayerOptions` to allow you set options like `playerColor` for the player